requirements.txt
fastapi==0.101
pydantic<2.5
typing-extensions==4.5.0
python-multipart
uvicorn
sqlalchemy==2.0.24
python-jose==3.3.0
passlib==1.7.4
python-dotenv
psycopg2-binary